app/requirements.txt
asgiref==3.8.1
click==8.1.7
Django==5.0.4
django-appconf==1.0.6
django-cors-headers==4.3.1
django-select2==8.1.2
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
geographiclib==2.0
geopy==2.4.1
gunicorn==22.0.0
h11==0.14.0
packaging==24.0
psycopg2-binary==2.9.9
PyJWT==2.8.0
sqlparse==0.5.0
uvicorn==0.29.0
